More crypto stuff

+++ Hap/crypt/curve25519.c
@@ -54,7 +54,7 @@ static void freeze(unsigned int a[32])
 
   for (j = 0;j < 32;++j) aorig[j] = a[j];
   add(a,a,minusp);
-  negative = -((a[31] >> 7) & 1);
+  negative = -(int)((a[31] >> 7) & 1);
   for (j = 0;j < 32;++j) a[j] ^= negative & (aorig[j] ^ a[j]);
 }
